+++ bob-common/src/configs/node.rs
@@ -270,6 +270,8 @@ pub struct Pearl {
     disks_events_logfile: String,
     #[serde(default)]
     bloom_filter_max_buf_bits_count: Option<usize>,
+    #[serde(default = "Pearl::default_validate_data_checksum_during_index_regen")]
+    validate_data_checksum_during_index_regen: bool,
 }
 
 impl Pearl {
@@ -290,6 +292,14 @@ impl Pearl {
             .parse::<HumanDuration>()
             .expect("parse humantime duration")
             .into()
+    }
+
+    fn default_validate_data_checksum_during_index_regen() -> bool {
+        false
+    }
+
+    pub fn validate_data_checksum_during_index_regen(&self) -> bool {
+        self.validate_data_checksum_during_index_regen
     }
 
     fn default_fail_retry_count() -> u64 {