+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Bob versions changelog
 - Client metrics are initialized at the start (#761)
 - Exist test on doubled range of keys for integration tests (#764)
 - Used swap and bob virtual memory metrics added to hardware metrics (#771)
+- Added validate_data_checksum_in_index_regen field to pearl config (#607)
 
 #### Changed
 - Make local put parallel to remote (#573)