+++ CHANGELOG.md
@@ -9,17 +9,22 @@ Bob versions changelog
 - Fast ping at the start (#657)
 - Client metrics are initialized at the start (#761)
 - Exist test on doubled range of keys for integration tests (#764)
+- Used swap and bob virtual memory metrics added to hardware metrics (#771)
+- Added validate_data_checksum_in_index_regen field to pearl config (#607)
 
 #### Changed
 - Make local put parallel to remote (#573)
 - Prefer online nodes for aliens, while maintaining uniform distribution (#571)
 - Build release binaries and docker images with `release-lto` profile (#714)
+- Use read lock instead of write on hierarchical filters update to improve performance (#596)
 
 #### Fixed
 - Fix incorrect exist result due to variables sharing between keys (#762)
+- Fix unit of measurement of memory in hardware metrics (#772)
 
 #### Updated
 - Logger to logstash updated to qoollo-log4rs-logstash v0.2 (#681)
+- Pearl updated to v0.18.0 (#778)
 
 
 ## [2.1.0-alpha.9] - 2023-01-16